backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
pandas==2.1.3
numpy==1.26.2
djangorestframework-simplejwt==5.3.0
reportlab==4.0.7
Pillow==10.1.0